+++ src/interfaces/four_momentum.jl
@@ -3,7 +3,7 @@
 
 Abstract base type for four-momentas, representing one energy and three spacial components.
 
-Also see: `QEDcore.SFourMomentum`, `QEDcore.MFourMomentum`
+Also see: [`QEDcore.SFourMomentum`](@extref), [`QEDcore.MFourMomentum`](@extref)
 """
 abstract type AbstractFourMomentum <: AbstractLorentzVector{Float64} end