+++ test/interfaces/process.jl
@@ -34,6 +34,8 @@ include("../test_implementation/TestImplementation.jl")
         @testset "failed process interface" begin
             @test_throws MethodError incoming_particles(TESTPROC_FAIL_ALL)
             @test_throws MethodError outgoing_particles(TESTPROC_FAIL_ALL)
+            @test_throws MethodError incoming_spin_pols(TESTPROC_FAIL_ALL)
+            @test_throws MethodError outgoing_spin_pols(TESTPROC_FAIL_ALL)
         end
 
         @testset "$PROC $MODEL" for (PROC, MODEL) in Iterators.product(
@@ -81,6 +83,31 @@ include("../test_implementation/TestImplementation.jl")
 
             @test number_particles(TESTPROC, dir, species) == groundtruth_particle_count
             @test number_particles(TESTPROC, test_ps) == groundtruth_particle_count
+        end
+    end
+
+    @testset "incoming/outgoing spins and polarizations" begin
+        groundtruth_incoming_spin_pols = ntuple(
+            x -> is_fermion(INCOMING_PARTICLES[x]) ? AllSpin() : AllPolarization(),
+            N_INCOMING,
+        )
+        groundtruth_outgoing_spin_pols = ntuple(
+            x -> is_fermion(OUTGOING_PARTICLES[x]) ? AllSpin() : AllPolarization(),
+            N_OUTGOING,
+        )
+
+        @test incoming_spin_pols(TESTPROC) == groundtruth_incoming_spin_pols
+        @test outgoing_spin_pols(TESTPROC) == groundtruth_outgoing_spin_pols
+        @test spin_pols(TESTPROC, Incoming()) == groundtruth_incoming_spin_pols
+        @test spin_pols(TESTPROC, Outgoing()) == groundtruth_outgoing_spin_pols
+
+        for (pt, sp) in Iterators.flatten((
+            Iterators.zip(incoming_particles(TESTPROC), incoming_spin_pols(TESTPROC)),
+            Iterators.zip(outgoing_particles(TESTPROC), outgoing_spin_pols(TESTPROC)),
+        ))
+            @test is_boson(pt) ? sp isa AbstractPolarization : true
+            @test is_fermion(pt) ? sp isa AbstractSpin : true
+            @test is_boson(pt) || is_fermion(pt)
         end
     end
 
@@ -175,5 +202,89 @@ include("../test_implementation/TestImplementation.jl")
             PhaseSpacePoint(TESTPROC, TESTMODEL, TESTPSDEF, ps_in_coords, ps_out_coords)
         @test groundtruth_in_psp ==
             InPhaseSpacePoint(TESTPROC, TESTMODEL, TESTPSDEF, ps_in_coords)
+    end
+end
+
+@testset "Process Multiplicity" begin
+    boson = TestImplementation.TestParticleBoson()
+    fermion = TestImplementation.TestParticleFermion()
+
+    _mult(::AbstractDefinitePolarization) = 1
+    _mult(::AbstractDefiniteSpin) = 1
+    _mult(::AbstractIndefinitePolarization) = 2
+    _mult(::AbstractIndefiniteSpin) = 2
+
+    @testset "no synced spins/pols" begin # test all possible combinations for fermion+boson->fermion+boson processes, without synced
+        spins = (SpinUp(), SpinDown(), AllSpin())
+        pols = (PolX(), PolY(), AllPol())
+        for (p1, s1, p2, s2) in Iterators.product(pols, spins, pols, spins)
+            proc = TestImplementation.TestProcessSP(
+                (boson, fermion), (boson, fermion), (p1, s1), (p2, s2)
+            )
+            @test multiplicity(proc) == prod(_mult.((p1, s1, p2, s2)))
+            @test incoming_multiplicity(proc) == prod(_mult.((p1, s1)))
+            @test outgoing_multiplicity(proc) == prod(_mult.((p2, s2)))
+        end
+    end
+
+    # some special cases for synced spins and pols testing
+    for i in 1:4
+        # i synced bosons
+        proc = TestImplementation.TestProcessSP(
+            (ntuple(_ -> boson, i)..., fermion),
+            (boson, fermion),
+            (ntuple(_ -> SyncedPolarization(1), i)..., AllSpin()),
+            (AllPol(), AllSpin()),
+        )
+        @test multiplicity(proc) == 16
+        @test incoming_multiplicity(proc) == 4
+        @test outgoing_multiplicity(proc) == 4
+
+        proc = TestImplementation.TestProcessSP(
+            (boson, fermion),
+            (ntuple(_ -> boson, i)..., fermion),
+            (AllPol(), SpinDown()),
+            (ntuple(_ -> SyncedPolarization(1), i)..., AllSpin()),
+        )
+        @test multiplicity(proc) == 8
+        @test incoming_multiplicity(proc) == 2
+        @test outgoing_multiplicity(proc) == 4
+
+        for j in 1:4
+            # ... with j synced fermions
+            proc = TestImplementation.TestProcessSP(
+                (ntuple(_ -> boson, i)..., fermion),
+                (boson, ntuple(_ -> fermion, j)),
+                (ntuple(_ -> SyncedPolarization(1), i)..., SpinDown()),
+                (PolX(), ntuple(_ -> SyncedSpin(1), j)...),
+            )
+            @test multiplicity(proc) == 4
+            @test incoming_multiplicity(proc) == 2
+            @test outgoing_multiplicity(proc) == 2
+        end
+    end
+
+    @testset "multiple differing synced polarizations" begin
+        proc = TestImplementation.TestProcessSP(
+            (ntuple(_ -> boson, 2)..., fermion),
+            (ntuple(_ -> boson, 2)..., fermion),
+            (ntuple(_ -> SyncedPolarization(1), 2)..., SpinUp()),
+            (ntuple(_ -> SyncedPolarization(2), 2)..., AllSpin()),
+        )
+        @test multiplicity(proc) == 8
+        @test incoming_multiplicity(proc) == 2
+        @test outgoing_multiplicity(proc) == 4
+    end
+
+    @testset "synced polarization across in and out particles" begin
+        proc = TestImplementation.TestProcessSP(
+            (ntuple(_ -> boson, 2)..., fermion),
+            (ntuple(_ -> boson, 2)..., fermion),
+            (ntuple(i -> SyncedPolarization(i), 2)..., SpinUp()),
+            (ntuple(i -> SyncedPolarization(i), 2)..., SpinDown()),
+        )
+        @test multiplicity(proc) == 4
+        @test incoming_multiplicity(proc) == 4
+        @test outgoing_multiplicity(proc) == 4
     end
 end