+++ test/particle_properties.jl
@@ -1,6 +1,10 @@
 using QEDbase
+using QEDcore
 using StaticArrays
 using Random
+
+include("test_implementation/TestImplementation.jl")
+using .TestImplementation: TestParticleBoson, TestParticleFermion
 
 # test function to test scalar broadcasting
 test_broadcast(x::AbstractParticle) = x
@@ -16,22 +20,53 @@ test_broadcast(x::AbstractSpinOrPolarization) = x
 
     @testset "spins and polarization" begin
         @testset "$spin_or_pol" for spin_or_pol in (
-            SpinUp(), SpinDown(), AllSpin(), PolX(), PolY(), AllPol()
+            SpinUp(),
+            SpinDown(),
+            AllSpin(),
+            PolX(),
+            PolY(),
+            AllPol(),
+            SyncedSpin(1),
+            SyncedPolarization(1),
         )
             @test test_broadcast.(spin_or_pol) == spin_or_pol
         end
     end
 end
 
-@testset "multiplicity of spins or pols" begin
-    @testset "single" begin
-        @testset "$spin_or_pol" for spin_or_pol in (SpinUp(), SpinDown(), PolX(), PolY())
-            @test multiplicity(spin_or_pol) == 1
-        end
-    end
-    @testset "multiple" begin
-        @testset "$spin_or_pol" for spin_or_pol in (AllSpin(), AllPol())
-            @test multiplicity(spin_or_pol) == 2
+TESTPROCS = (
+    TestImplementation.TestProcessSP(
+        (TestParticleBoson(), TestParticleFermion()),
+        (TestParticleBoson(), TestParticleFermion()),
+        (AllPol(), AllSpin()),
+        (AllPol(), AllSpin()),
+    ),
+    TestImplementation.TestProcessSP(
+        (TestParticleBoson(), TestParticleBoson(), TestParticleFermion()),
+        (TestParticleBoson(), TestParticleFermion()),
+        (SyncedPol(1), SyncedPol(1), AllSpin()),
+        (AllPol(), AllSpin()),
+    ),
+    TestImplementation.TestProcessSP(
+        (TestParticleBoson(), TestParticleBoson(), TestParticleFermion()),
+        (TestParticleBoson(), TestParticleFermion()),
+        (SyncedPol(1), SyncedPol(2), SyncedSpin(2)),
+        (SyncedPol(2), SyncedSpin(2)),
+    ),
+)
+
+@testset "spin_pol iterator ($proc)" for proc in TESTPROCS
+    @test length(spin_pols_iter(proc)) == multiplicity(proc)
+
+    for combinations in spin_pols_iter(proc)
+        @test length(combinations) == 2
+        in_comb, out_comb = combinations
+
+        @test length(in_comb) == length(incoming_particles(proc))
+        @test length(out_comb) == length(outgoing_particles(proc))
+
+        for sp in Iterators.flatten((in_comb, out_comb))
+            @test sp isa AbstractDefiniteSpin || sp isa AbstractDefinitePolarization
         end
     end
 end