+++ test/test_implementation/test_process.jl
@@ -22,6 +22,24 @@ end
 QEDbase.incoming_particles(proc::TestProcess) = proc.incoming_particles
 QEDbase.outgoing_particles(proc::TestProcess) = proc.outgoing_particles
 
+"""
+    TestProcessSP
+
+Process for testing with settable spin and polarization.
+"""
+struct TestProcessSP{IP<:Tuple,OP<:Tuple,IN_SP<:Tuple,OUT_SP<:Tuple} <:
+       AbstractProcessDefinition
+    incoming_particles::IP
+    outgoing_particles::OP
+    incoming_spin_pols::IN_SP
+    outgoing_spin_pols::OUT_SP
+end
+
+QEDbase.incoming_particles(proc::TestProcessSP) = proc.incoming_particles
+QEDbase.outgoing_particles(proc::TestProcessSP) = proc.outgoing_particles
+QEDbase.incoming_spin_pols(proc::TestProcessSP) = proc.incoming_spin_pols
+QEDbase.outgoing_spin_pols(proc::TestProcessSP) = proc.outgoing_spin_pols
+
 struct TestProcess_FAIL{IP<:Tuple,OP<:Tuple} <: AbstractProcessDefinition
     incoming_particles::IP
     outgoing_particles::OP
@@ -55,7 +73,7 @@ function TestProcess_FAIL_ALL(rng::AbstractRNG, N_in::Int, N_out::Int)
 end
 
 """
-Test process with no implemented interface except the incoming and outgoing particles. 
+Test process with no implemented interface except the incoming and outgoing particles.
 Should fail every usage except construction of itself and the respective phase space point for given four-momenta.
 """
 struct TestProcess_FAIL_DIFFCS{IP<:Tuple,OP<:Tuple} <: AbstractProcessDefinition