+++ test/test_implementation/utils.jl
@@ -1,5 +1,5 @@
 
-# Check if any failed type is in the input 
+# Check if any failed type is in the input
 _any_fail(x...) = true
 _any_fail(::TestProcess, ::TestModel) = false
 _any_fail(::TestProcess, ::TestModel, ::TestPhasespaceDef) = false